+++ packages/fuels-core/src/types.rs
@@ -14,6 +14,7 @@ pub mod errors;
 pub mod param_types;
 mod token;
 pub mod transaction_builders;
+pub mod tx_response;
 pub mod tx_status;
 mod wrappers;
 pub use dry_runner::*;
