+++ packages/fuels-core/src/types/tx_status.rs
@@ -6,31 +6,14 @@ use fuel_abi_types::error_codes::{
 use fuel_core_client::client::types::TransactionStatus as ClientTransactionStatus;
 #[cfg(feature = "std")]
 use fuel_core_types::services::executor::{TransactionExecutionResult, TransactionExecutionStatus};
-use fuel_tx::{Receipt, TxId};
+use fuel_tx::Receipt;
 #[cfg(feature = "std")]
 use fuel_vm::state::ProgramState;
 
 use crate::{
     codec::LogDecoder,
-    sealed::Sealed,
     types::errors::{transaction::Reason, Error, Result},
 };
-
-impl Sealed for TxId {}
-impl Sealed for Option<TxId> {}
-
-pub trait TxResponseType: Sealed {}
-
-impl TxResponseType for TxId {}
-impl TxResponseType for Option<TxId> {}
-
-#[derive(Clone, Debug)]
-pub struct TxResponse<T: TxResponseType = TxId> {
-    pub receipts: Vec<Receipt>,
-    pub gas_used: u64,
-    pub total_fee: u64,
-    pub id: T,
-}
 
 #[derive(Debug, Clone)]
 pub enum TxStatus {
